Make LEDs dynamically settable based on a value.
led_data now has a type, either REG or DYN.  When type is REG, member reg
contains the same data as before and the LEDs are updated in the same manner.
Attributes led_logo and leds_ring set the LED data to type REG, and work exactly
as before.

The new DYN type has a value source (currently only liquid temperature is
implemented) based on which it sets the LEDs on each update, if necessary.  The
new *_dyn attributes set the LED data to type DYN.

Also use #define-constants, sizeof, and ARRAY_SIZE in a more consistent manner.

+++ src/kraken_x62/percent.h
@@ -5,7 +5,7 @@
 
 #include <linux/mutex.h>
 
-#define PERCENT_MSG_SIZE 5
+#define PERCENT_MSG_SIZE ((size_t) 5)
 
 struct percent_data {
 	u8 msg[PERCENT_MSG_SIZE];
